move populateCatalogObject back into bucketService
