+++ src/main/resources/application.properties
@@ -38,7 +38,6 @@ hibernate.current_session_context_class=thread
 spring.jpa.properties.hibernate.implicit_naming_strategy=org.hibernate.boot.model.naming.ImplicitNamingStrategyJpaCompliantImpl
 spring.jpa.properties.hibernate.ejb.naming_strategy_delegator=
 spring.jpa.properties.hibernate.id.new_generator_mappings=false
-spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 
 # Show or not log for each sql query
 spring.jpa.show-sql=false
